fix: make timestamp an int

+++ src/ape/api/providers.py
@@ -338,7 +338,7 @@ class BlockAPI:
     number: int
     parent_hash: HexBytes
     size: int
-    timestamp: float
+    timestamp: int
 
     @classmethod
     @abstractmethod
